Update E2E - check personal details fields; simplify stub data
