+++ e2e_tests/integration/stepDefinitions/index.ts
@@ -1,6 +1,6 @@
 import { defineStep, When } from '@badeball/cypress-cucumber-preprocessor'
-import { DateTime } from 'luxon'
-import { formatObjectDate, getTestDataPerEnvironment, isoDateToObject } from '../../utils'
+import { getTestDataPerEnvironment } from '../../utils'
+import { longDateMatchPattern, formatIsoDate } from '../../../cypress_shared/utils'
 
 const apiDataForCrn = getTestDataPerEnvironment()
 
@@ -190,6 +190,12 @@ defineStep('Maria confirms {string} to a risk of contraband', (answer: string) =
 When('Maria reviews the personal details', () => {
   cy.getElement('Personal details To review').should('exist')
   cy.clickLink('Personal details')
+  cy.getText('fullName').as('fullName')
+  cy.getDefinitionListValue('Name').should('match', apiDataForCrn.fullName)
+  cy.getText('gender').as('gender')
+  cy.getDefinitionListValue('Gender').should('match', apiDataForCrn.gender)
+  cy.getDateAttribute('dateOfBirth').as('dateOfBirth')
+  cy.getDefinitionListValue('Date of birth').should('match', longDateMatchPattern(apiDataForCrn.dateOfBirth))
   cy.clickLink('Continue')
   cy.getElement('Personal details Reviewed').should('exist')
 })
@@ -212,11 +218,10 @@ When('Maria enters the offence analysis', () => {
 When('Maria enters the previous releases', () => {
   cy.getElement('Previous releases To do').should('exist')
   cy.clickLink('Previous releases')
+  cy.getDateAttribute('lastReleaseDate').as('lastReleaseDate')
   cy.get('@offenderName').then(offenderName => {
     cy.selectRadio(`Has ${offenderName} been released previously?`, 'Yes')
   })
-  // store last release date so it can be checked for later in the Part A
-  cy.get(`[data-qa="lastReleaseDate"]`).invoke('attr', 'data-date').as('lastReleaseDate')
   cy.clickButton('Continue')
   cy.enterDateTime(apiDataForCrn.previousReleaseDate)
   cy.clickButton('Continue')
@@ -284,72 +289,3 @@ When('Maria confirms the person is in prison custody', () => {
 When('Maria confirms the existing custody status', () => {
   cy.clickButton('Continue')
 })
-
-export const q1EmergencyRecall = (contents: string, answer: string) =>
-  expect(contents).to.contain(`until PPCS has issued the revocation order.  ${answer}`)
-export const q2IndeterminateSentenceType = (contents: string, answer: string) =>
-  expect(contents).to.contain(`Is the offender serving a life or IPP/DPP sentence? ${answer}`)
-export const q3ExtendedSentence = (contents: string, answer: string) =>
-  expect(contents).to.contain(`Is the offender serving one of the following:  ${answer}`)
-export const q4OffenderDetails = (contents: string) => {
-  expect(contents).to.match(apiDataForCrn.fullName as RegExp)
-  expect(contents).to.match(apiDataForCrn.dateOfBirth as RegExp)
-  expect(contents).to.match(apiDataForCrn.ethnicity as RegExp)
-  expect(contents).to.match(apiDataForCrn.gender as RegExp)
-  expect(contents).to.match(apiDataForCrn.cro as RegExp)
-  expect(contents).to.match(apiDataForCrn.pnc as RegExp)
-  expect(contents).to.match(apiDataForCrn.prisonNo as RegExp)
-  expect(contents).to.match(apiDataForCrn.noms as RegExp)
-
-  cy.get('@lastReleaseDate').then(lastReleaseDate => {
-    const lastReleaseDateFormatted = formatObjectDate(isoDateToObject(lastReleaseDate))
-    const previousReleaseDateFormatted = formatObjectDate(apiDataForCrn.previousReleaseDate)
-    expect(contents).to.contain(
-      `Date of last release and previous release: ${lastReleaseDateFormatted}, ${previousReleaseDateFormatted}`
-    )
-  })
-}
-export const q5SentenceDetails = (contents: string) => {
-  expect(contents).to.match(apiDataForCrn.indexOffence as RegExp)
-  expect(contents).to.match(apiDataForCrn.dateOfOriginalOffence as RegExp)
-  expect(contents).to.match(apiDataForCrn.dateOfSentence as RegExp)
-  expect(contents).to.match(apiDataForCrn.lengthOfSentence as RegExp)
-  expect(contents).to.match(apiDataForCrn.licenceExpiryDate as RegExp)
-  expect(contents).to.match(apiDataForCrn.sentenceExpiryDate as RegExp)
-  expect(contents).to.match(apiDataForCrn.custodialTerm as RegExp)
-  expect(contents).to.match(apiDataForCrn.extendedTerm as RegExp)
-}
-export const q6CustodyStatus = (contents: string, answer: string) =>
-  expect(contents).to.contain(`Is the offender currently in police custody or prison custody? ${answer}`)
-export const q7Addresses = (contents: string, answer: string) =>
-  expect(contents).to.contain(`If the offender is in police custody, state where: ${answer}`)
-export const q8ArrestIssues = (contents: string, answer: string, details: string) => {
-  expect(contents).to.contain(`Are there any arrest issues of which police should be aware?  ${answer}`)
-  expect(contents).to.contain(
-    `If yes, provide details below, including information about any children or vulnerable adults linked to any of the above addresses: ${details}`
-  )
-}
-export const q12MappaDetails = (contents: string) => {
-  expect(contents).to.match(apiDataForCrn.mappaCategory as RegExp)
-  expect(contents).to.match(apiDataForCrn.mappaLevel as RegExp)
-}
-export const q16IndexOffenceDetails = (contents: string, offenceAnalysis?: string) => {
-  if (offenceAnalysis) {
-    expect(contents).to.contain(offenceAnalysis)
-  } else {
-    expect(contents).to.match(apiDataForCrn.indexOffenceDetails as RegExp)
-  }
-}
-
-export const q22RecallType = (contents: string, answer: string, details: string) => {
-  expect(contents).to.contain(`Select the proposed recall type, having considered the information above: ${answer}`)
-  expect(contents).to.contain(`Explain your reasons for the above recall type recommendation: ${details}`)
-}
-export const q25ProbationDetails = (contents: string) => {
-  expect(contents).to.match(apiDataForCrn.nameOfPersonCompletingForm as RegExp)
-  expect(contents).to.match(apiDataForCrn.emailAddressOfPersonCompletingForm as RegExp)
-  expect(contents).to.match(apiDataForCrn.region as RegExp)
-  expect(contents).to.match(apiDataForCrn.ldu as RegExp)
-  expect(contents).to.contain(`${apiDataForCrn.dateOfDecision} ${DateTime.now().toFormat('dd/MM/y')}`)
-  expect(contents).to.match(apiDataForCrn.timeOfDecision as RegExp)
-}
